+++ flamingo/example/lib/model/user.flamingo.dart
@@ -9,6 +9,8 @@ part of 'user.dart';
 /// Field value key
 enum UserKey {
   name,
+  editAt,
+  clearAt,
 
   setting,
 }
@@ -18,6 +20,10 @@ extension UserKeyExtension on UserKey {
     switch (this) {
       case UserKey.name:
         return 'name';
+      case UserKey.editAt:
+        return 'editAt';
+      case UserKey.clearAt:
+        return 'clearAt';
       case UserKey.setting:
         return 'setting';
       default:
@@ -30,6 +36,8 @@ extension UserKeyExtension on UserKey {
 Map<String, dynamic> _$toData(User doc) {
   final data = <String, dynamic>{};
   Helper.writeNotNull(data, 'name', doc.name);
+  Helper.writeNotNull(data, 'editAt', doc.editAt);
+  Helper.writeNotNull(data, 'clearAt', doc.clearAt);
 
   return data;
 }
@@ -37,4 +45,15 @@ Map<String, dynamic> _$toData(User doc) {
 /// For load data
 void _$fromData(User doc, Map<String, dynamic> data) {
   doc.name = Helper.valueFromKey<String>(data, 'name');
+  if (data['editAt'] is Map) {
+    doc.editAt = Helper.timestampFromMap(data, 'editAt');
+  } else {
+    doc.editAt = Helper.valueFromKey<Timestamp>(data, 'editAt');
+  }
+
+  if (data['clearAt'] is Map) {
+    doc.clearAt = Helper.timestampFromMap(data, 'clearAt');
+  } else {
+    doc.clearAt = Helper.valueFromKey<Timestamp>(data, 'clearAt');
+  }
 }