+++ flamingo_generator/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.6.0
+Modified Timestamp type.
+
+## 0.5.0
+Modified Increment field.
+
 # 0.4.1+1
 Modified homepage url.